+++ .env.example
@@ -12,3 +12,6 @@ REACT_APP_COLLECTION_IMAGE=collection/collection-example-image.png
 REACT_APP_RPC_URL_1=
 # Goerli
 REACT_APP_RPC_URL_5=
+
+# Wallet connect project id to allow users to use WalletConnect as provider
+REACT_APP_WALLET_CONNECT_PROJECT_ID=
